refactor: mock data 파일명 수정 - mockData -> data
